refactor: reorganize tests code
