+++ tests/unit/cases/Helpers/FunctionsTest.php
