+++ tests/unit/cases/Helpers/WpHooksTest.php
