+++ tests/unit/fixtures/argument-type-declaration.php
