+++ tests/unit/fixtures/disallow-magic-serialize.php
