+++ tests/unit/fixtures/disallow-short-open-tag.php
