+++ tests/unit/fixtures/encoding-comment.php
