+++ tests/unit/fixtures/forbidden-public-property.php
