+++ tests/unit/fixtures/function-body-start.php
