+++ tests/unit/fixtures/function-length-no-blank-lines.php
