+++ tests/unit/fixtures/hook-closure-return.php
