+++ tests/unit/fixtures/no-accessors.php
