+++ tests/unit/fixtures/no-root-namespace-functions-multi.php
