+++ tests/unit/fixtures/no-root-namespace-functions-single.php
