+++ tests/unit/fixtures/no-top-level-define.php
