+++ tests/unit/fixtures/property-per-class-limit.php
