+++ tests/unit/fixtures/return-type-declaration.php
