+++ tests/unit/fixtures/static-closure.php
