+++ tests/unit/fixtures/trailing-semicolon.php
