+++ tests/unit/fixtures/var-names-camel-case.php
