+++ tests/unit/fixtures/var-names-no-local.php
