+++ tests/unit/fixtures/var-names-no-props.php
